ImageUtils refactored, metanorma/metanorma-jis#226
